파일명 변경 및 전역함수 수정, 파일 분리 반영
