+++ Meditory/Meditory/Network/DTO/CSEResponse.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct CSEResponse: Decodable {
+  let items: [CSEItem]?
+}
